Folded MapContainerSupport into MapContainer
